feat(eo): add EoCertificateType enum to domain
This commit adds the `EoCertificateType` enum to the `EoCertificate` domain in order to provide a more structured and type-safe representation of the certificate types. This enum includes the values `Production` and `Consumption` to replace the previous string literals used for `certificateType`. This change improves code readability and reduces the risk of introducing errors when working with certificate types.