feat(eo, metering points): add helper modal (#3411)
* feat(eo, metering points): add helper modal

* style: format

* feat(eo, metering points): adjust translations

* feat(eo, metering points): adjust translations

* fix(eo, certificates): pagination

* style: format

* trigger ci

* feat(eo): add EoCertificateType enum to domain

This commit adds the `EoCertificateType` enum to the `EoCertificate` domain in order to provide a more structured and type-safe representation of the certificate types. This enum includes the values `Production` and `Consumption` to replace the previous string literals used for `certificateType`. This change improves code readability and reduces the risk of introducing errors when working with certificate types.

* tmp: change certificates per page to test pagination

* refactor(eo, certificates): use correct loading

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>