Do not wrap content in dh-result
This causes "waterfall rendering", meaning that tables and other data
cannot start loading until GetMeteringPoint has returned, even when
other components do not depend on this data. Loading and error state
will come later.

The reason for this, beside the faster rendering, is that the drawer had
issues with its initial sizing when it was within an async router-outlet
(watt-link-tabs). The width would be 0, unless I changed the page to
"eagerly" load children.